fix_calendar_secret_id_type

+++ db/migrate/20200629062107_add_column_to_store.rb
@@ -1,5 +1,5 @@
 class AddColumnToStore < ActiveRecord::Migration[6.0]
   def change
-    add_column :stores, :calendar_secret_id, :bigint
+    add_column :stores, :calendar_secret_id, :integer
   end
 end
